Bash client scripts and zip-archive builder created

+++ windows-client/scripts/local-processing.sh
@@ -1,27 +1,76 @@
 #!/bin/bash
+## Run the local processing (datasource.xml to fhir and pseudonymisation)
 
 df="%Y-%m-%d %H:%M:%S"
-function log_debug { [[ "${log_verbosity}" -ge 3 ]] && >&2 echo "[$(date +"$df")] DEBUG: ${@}"; }
-function log_info { [[ "${log_verbosity}" -ge 2 ]] && >&2 echo "[$(date +"$df")] INFO: ${@}"; }
-function log_warn { [[ "${log_verbosity}" -ge 1 ]] && >&2 echo "[$(date +"$df")] WARN: ${@}"; }
-function log_error { >&2 echo "[$(date +"$df")] ERROR: ${@}"; }
+script_name="$(basename ${BASH_SOURCE[0]})"
+function log_debug { [[ "${log_verbosity}" -ge 3 ]] && >&2 echo "[$(date +"$df")] {${script_name}} DEBUG: ${@}"; }
+function log_info { [[ "${log_verbosity}" -ge 2 ]] && >&2 echo "[$(date +"$df")] {${script_name}} INFO: ${@}"; }
+function log_warn { [[ "${log_verbosity}" -ge 1 ]] && >&2 echo "[$(date +"$df")] {${script_name}} WARN: ${@}"; }
+function log_error { >&2 echo "[$(date +"$df")] {${script_name}} ERROR: ${@}"; }
 >&2 echo "(stderr) logging verbosity set to: ${log_verbosity}"
 
-## Installer for git on windows/git-bash
-[[ java -version > /dev/null 2>&1 ]] || log_error "Java not available" && exit 1
-[[ python --version > /dev/null 2>&1 ]] || log_error "Python not available" && exit 1
+## Add the venv python to PATH, so we use that over the system version (if it was created by install.sh)
+client_basedir=$( dirname $(dirname "$(readlink -f "${BASH_SOURCE[0]}")") ) ## Parent of this "scripts" dir
+[[ -d "${client_basedir}/.venv" ]] && source .venv/bin/activate && log_info "Using python venv..."
+
+## Check java and python are available
+java -version > /dev/null 2>&1 || (echo "Java not available" && exit 1)
+python --version > /dev/null 2>&1 || (echo "Python not available" && exit 1)
+
+## See if the user wants to restrict to -s (single-stage) "f" (fhir) or "p" (pseudonym) processing
+single_stage=
+SOPTS="s:"
+TMP=$(getopt -o "$SOPTS" -n "$CMD" -- "$@") || exit 1
+eval set -- "$TMP"
+unset TMP
+while true; do
+    case "$1" in
+        -s)
+            single_stage="$2"
+            shift
+            ;;
+        --)                                       # end of options
+            shift
+            break
+            ;;
+    esac
+    shift
+done
+if [[ $# != 1 ]] ; then
+    log_error "You must provide exactly 1 argument, the path to the 'datasource.xml' file"
+    exit 1
+fi
+## Check secret_key variable - disallow default "ChangeMe" and prompt if missing/empty
+if [[ ! -n $secret_key || $secret_key = "ChangeMe" ]] ; then
+    log_info "secret_key is default or not set, prompting for user input..."
+    while true; do
+        read -p "Please enter your secret_key: " api_key
+        ## Do some sort of sanity checks on key
+        if [[ "$api_key" =~ ^[A-Za-z][A-Za-z0-9\`\&\;\'\<\>_#$%@^~*+!?=.,:-]*$ ]] ; then
+            secret_key=$api_key
+            break
+        else
+            echo "Sorry, this name isn't valid, please try again..." >&2
+        fi
+    done
+fi
+
+## Set the datasource_path and _dir for both processing parts to use
+datasource_path=$1
+[[ -f "${datasource_path}" ]] || (log_error "Provided parameter (${datasource_path}) is not a file on this filesystem, aborting..." && exit 1)
+datasource_dir=$( dirname "${datasource_path}" )
+
 
 function datasource_to_fhir {
     ## Prepare the source data into a pseudonymised fhir-bundle
-    datasource_path=$1
-    [[ -f ${datasource_path} ]] || echo "File doesn't exist! ${datasource_path}" && exit 1
+    [[ -f ${datasource_path} ]] || (echo "File doesn't exist! ${datasource_path}" && exit 1)
     log_info "Processing a specific source (${datasource_path})..."
-    datasource_dir=$( dirname "${datasource_path}" )
-    mkdir "${datasource_dir}/client-output/"
-    rm ${datasource_dir}/client-output/*
+    log_debug "Checking dir is ready: ${datasource_dir}"
+    [[ -d ${datasource_dir}/client-output ]] && rm ${datasource_dir}/client-output/* || mkdir "${datasource_dir}/client-output/"
     raw_fhir_path=${datasource_dir}/client-output/fhir-bundle-raw.xml
-    fhir-from-datasource.sh ${datasource_path} > ${raw_fhir_path}
-    log_debug "Data now in fhir-bundle format: ${raw_fhir_path}"
+    log_debug "Directories and paths set, processing... (${client_basedir}/format/fhir-from-datasource.sh ${datasource_path})"
+    ${client_basedir}/format/fhir-from-datasource.sh ${datasource_path} > ${raw_fhir_path}
+    log_info "Data now in fhir-bundle format: ${raw_fhir_path}"
 }
 
 function pseudonymise_fhir {
@@ -30,10 +79,21 @@ function pseudonymise_fhir {
     ## The python pseudonymisation script uses env vars for settings
     export input_fn=${datasource_dir}/client-output/fhir-bundle-raw.xml
     export output_fn=${datasource_dir}/client-output/fhir-bundle-dwh.xml
-    export secret_key=${datasource_dir}/client-output/fhir-bundle-dwh.xml
-    pseudonym-pid-fhir.py
-    log_info "Processing complete, see DWH-ready output at: (${output_fn})..."
+    export secret_key=${secret_key}
+    ${client_basedir}/pseudonym/src/pseudonym-pid-fhir.py
+    log_info "Pseudonym processing complete, see DWH-ready output at: (${output_fn})..."
 }
 
-datasource_to_fhir $1
-pseudonymise_fhir
+## START
+if [[ ${single_stage} = "f" ]] ; then
+    datasource_to_fhir
+elif [[ ${single_stage} = "p" ]] ; then
+    pseudonymise_fhir
+else
+    datasource_to_fhir
+    pseudonymise_fhir
+fi
+## END
+
+## Cleanup venv...
+deactivate
